requirements.txt
amqp==2.3.2
Click==7.0
dnspython==1.15.0
eventlet==0.24.1
Flask==1.0.2
Flask-SocketIO==3.0.2
greenlet==0.4.15
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.1.0
monotonic==1.5
python-engineio==2.3.2
python-socketio==2.0.0
redis==2.10.6
six==1.11.0
vine==1.1.4
Werkzeug==0.14.1
psycopg2-binary==2.7.6.1
